Move all the operators into a separate subpackage.
