+++ core/src/main/kotlin/org/nwapw/abacus/plugin/standard/operator/OperatorNegate.kt
@@ -1,4 +1,4 @@
-package org.nwapw.abacus.plugin.standard
+package org.nwapw.abacus.plugin.standard.operator
 
 import org.nwapw.abacus.context.MutableEvaluationContext
 import org.nwapw.abacus.function.NumberOperator
@@ -7,15 +7,16 @@ import org.nwapw.abacus.function.OperatorType
 import org.nwapw.abacus.number.NumberInterface
 
 /**
- * The division operator.
+ * The negation operator.
  *
- * This is a standard operator that simply performs division.
+ * This is a standard operator that negates a number.
  */
-class OperatorDivide: NumberOperator(OperatorAssociativity.LEFT, OperatorType.BINARY_INFIX, 1) {
+class OperatorNegate: NumberOperator(OperatorAssociativity.LEFT, OperatorType.UNARY_PREFIX, 0) {
 
     override fun matchesParams(context: MutableEvaluationContext, params: Array<out NumberInterface>) =
-            params.size == 2
+            params.size == 1
+
     override fun applyInternal(context: MutableEvaluationContext, params: Array<out NumberInterface>) =
-            params[0] / params[1]
+            -params[0]
 
 }
