+++ core/src/main/kotlin/org/nwapw/abacus/plugin/standard/operator/OperatorSubtract.kt
@@ -1,4 +1,4 @@
-package org.nwapw.abacus.plugin.standard
+package org.nwapw.abacus.plugin.standard.operator
 
 import org.nwapw.abacus.context.MutableEvaluationContext
 import org.nwapw.abacus.function.NumberOperator
@@ -7,16 +7,15 @@ import org.nwapw.abacus.function.OperatorType
 import org.nwapw.abacus.number.NumberInterface
 
 /**
- * The negation operator.
+ * The subtraction operator.
  *
- * This is a standard operator that negates a number.
+ * This is a standard operator that performs subtraction.
  */
-class OperatorNegate: NumberOperator(OperatorAssociativity.LEFT, OperatorType.UNARY_PREFIX, 0) {
+class OperatorSubtract: NumberOperator(OperatorAssociativity.LEFT, OperatorType.BINARY_INFIX, 0) {
 
     override fun matchesParams(context: MutableEvaluationContext, params: Array<out NumberInterface>) =
-            params.size == 1
-
+            params.size == 2
     override fun applyInternal(context: MutableEvaluationContext, params: Array<out NumberInterface>) =
-            -params[0]
+            params[0] - params[1]
 
 }
